anadidas reglas para versiones antiguas de polkit

+++ installer.sh
@@ -8,4 +8,7 @@ cp 1?-usb*.rules /etc/udev/rules.d/
 udevadm control --reload-rules
 
 #Copy polkit rules
-cp 10-*shutdown.rules /usr/share/polkit-1/rules.d/
+if test $(pkaction --version | cut -d " " -f 3 | cut -d "." -f 2) -ge 106
+then
+	cp 10-*shutdown.rules /usr/share/polkit-1/rules.d/
+fi
